requirements.txt
# requirements.txt – versiones probadas con Python 3.12 (jun-25)

# ─── Async stack ──────────────────────────────────────────────
aiohttp>=3.9,<4
aiosqlite>=0.19          # driver async para SQLite
tenacity>=8.2            # reintentos con back-off

# ─── ORM & BD ─────────────────────────────────────────────────
SQLAlchemy[asyncio]>=2.0.30   # incluye extensión async
python-dotenv>=1.0            # carga .env

# ─── Datos / indicadores ─────────────────────────────────────
pandas>=2.2
numpy>=2.0                   # requerido por pandas
ta>=0.11                     # EMA/RSI puro-python
pytz>=2024.1

# ─── Solana signing ───────────────────────────────────────────
solders>=0.15                # keypair & tx (binding Rust)
solana>=0.32
base58>=2.1

# ─── Extra utilidades ─────────────────────────────────────────
typing-extensions>=4.11       # para hints futuros
